Added SSH Communication Method for using RPI4 as trigger mechanism. Need to implement SFTP method to download files.
Also need to prevent photo capture, start rapid when subsystems are not running. these result in fatal errors because they make calls to the subsystem

Alot of Cleanup to do After this round of edits. Need to add comments to all code.

+++ PhotoInspect/Photogrametry/gphoto_CTRL.cs
@@ -1,10 +1,13 @@
-﻿using System;
+﻿using Renci.SshNet;
+using Renci.SshNet.Common;
+using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Text.RegularExpressions;
+using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
@@ -17,6 +20,8 @@ namespace Photogrametry
         public RapidFunctions Rap;
         private Process WSLproc;
         private bool WSLRunning=false;
+        private SshClient sshCLI;
+        private ShellStream sshStream;
         public gphoto_CTRL(Form1 _form1)
         {
             this.fr = _form1;
@@ -24,59 +29,70 @@ namespace Photogrametry
         }
        // The below function starts the windows subsystem for linux Binds the camera hardware and returns
        // a process item that can be used to access the WSL from the application.
-        public void WSLStart() 
+        public void SubSystemStart() 
         {
             try
             {
-                
-                    
-                WSLproc = new Process
+
+            if (fr.chk_UseWSL.Checked)
                 {
-                    StartInfo = new ProcessStartInfo
+                    WSLproc = new Process
                     {
-                        FileName = @"wsl.exe",
-                        UseShellExecute = false,
-                        RedirectStandardOutput = true,
-                        RedirectStandardError = false,
-                        RedirectStandardInput = true,
-                        CreateNoWindow = true,
-                        //          Arguments = "gphoto2 --capture-image-and-download -F=" + num_to_cap + " -I=" + interval + " --force-overwrite",
+                        StartInfo = new ProcessStartInfo
+                        {
+                            FileName = @"wsl.exe",
+                            UseShellExecute = false,
+                            RedirectStandardOutput = true,
+                            RedirectStandardError = false,
+                            RedirectStandardInput = true,
+                            CreateNoWindow = true,
+                            // Arguments = "gphoto2 --capture-image-and-download -F=" + num_to_cap + " -I=" + interval + " --force-overwrite",
+                        }
+                    };
+
+                    WSLproc.Start();
+                    WSLproc.StandardInput.NewLine = "\n";
+                    if (!WSLproc.HasExited)
+                    {
+                        WSLRunning = true;
+                        fr.SubRunning.BackColor = System.Drawing.Color.Green;
+                        fr.SubRunning.Text = "Running";
                     }
 
-                };
+                    ProcessStartInfo startInfo = new ProcessStartInfo
+                    {
+                        FileName = "usbipd",
+                        Arguments = $"attach --wsl --hardware-id {fr.s_hardwareID}",
+                        RedirectStandardOutput = true,
+                        RedirectStandardError = true,
+                        UseShellExecute = false,
+                        CreateNoWindow = true,
+                    };
 
-                WSLproc.Start();
-                WSLproc.StandardInput.NewLine = "\n";
-                if (!WSLproc.HasExited)
-                {
-                    WSLRunning = true;
-                    fr.SubRunning.BackColor = System.Drawing.Color.Green;
-                    fr.SubRunning.Text = "Running";
+                    using (Process process = new Process { StartInfo = startInfo })
+                    {
+                        process.Start();
+                        process.WaitForExit();
+                    }
+
+                    // Create a new folder in the Home Directory for this session and go there for capturing
+                    WSLproc.StandardInput.WriteLine(@"cd ~/camera/");
+                    string now = DateTime.Now.ToString("dd_MM_yy_HH_mm");
+                    WSLproc.StandardInput.WriteLine($"mkdir photos_{now}");
+                    WSLproc.StandardInput.WriteLine($"cd photos_{now}");
                 }
-
-                ProcessStartInfo startInfo = new ProcessStartInfo
+                else
                 {
-                    FileName = "usbipd",
-                    Arguments = $"attach --wsl --hardware-id {fr.s_hardwareID}",
-                    RedirectStandardOutput = true,
-                    RedirectStandardError = true,
-                    UseShellExecute = false,
-                    CreateNoWindow = true,
-                    
-                };
-
-                using (Process process = new Process { StartInfo = startInfo })
-                {
-                    process.Start();
-                    process.WaitForExit();
+                    SSHStart();
+                    sshStream = sshCLI.CreateShellStream("input", 0, 0, 0, 0, 1000000);
+                 // Create a new folder in the Home Directory for this session and go there for capturing
+                 
+                    SSHWrite("mkdir camera");
+                    SSHWrite("cd camera");
+                    string now = DateTime.Now.ToString("dd_MM_yy_HH_mm");
+                    SSHWrite($"mkdir photos_{now}");
+                    SSHWrite($"cd photos_{now}");
                 }
-
-                // Create a new folder in the Home Directory for this session and go there for capturign
-                WSLproc.StandardInput.WriteLine(@"cd ~/camera/");
-                string now = DateTime.Now.ToString("dd_MM_yy_HH_mm");
-                WSLproc.StandardInput.WriteLine($"mkdir photos_{now}");
-                WSLproc.StandardInput.WriteLine($"cd photos_{now}");
-
             }
             catch(System.Exception ex) 
             {
@@ -87,39 +103,48 @@ namespace Photogrametry
         }
        
         //This Function kills the WSL system that was started and detaches the USB
-        public void WSLClose()
+        public void SubSystemStop()
         {
             try
             {
-                if (WSLRunning == true)
+                if(fr.chk_UseWSL.Checked)
                 {
-                    if (fr.DeleteFilesOnExit.Checked == true)
+                    if (WSLRunning == true)
                     {
-                        WSLproc.StandardInput.WriteLine($"sudo gphoto2 --delete-all-files -f={fr.CameraFolder.Text}");
+                        if (fr.DeleteFilesOnExit.Checked == true)
+                        {
+                            WSLproc.StandardInput.WriteLine($"sudo gphoto2 --delete-all-files -f={fr.CameraFolder.Text}");
+                        }
+                        WSLproc.Kill();
+                        WSLproc.Dispose();
+                        WSLRunning = false;
+                        fr.SubRunning.BackColor = System.Drawing.Color.Red;
+                        fr.SubRunning.Text = "Not Running";
                     }
-                    WSLproc.Kill();
-                    WSLproc.Dispose();
-                    WSLRunning=false;
-                    fr.SubRunning.BackColor = System.Drawing.Color.Red;
-                    fr.SubRunning.Text = "Not Running";
-                }
 
-                ProcessStartInfo startInfo = new ProcessStartInfo
-                {
-                    FileName = "usbipd",
-                    Arguments = $"detach --hardware-id {fr.s_hardwareID}",
-                    RedirectStandardOutput = true,
-                    RedirectStandardError = true,
-                    UseShellExecute = false,
-                    CreateNoWindow = true
-                };
+                    ProcessStartInfo startInfo = new ProcessStartInfo
+                    {
+                        FileName = "usbipd",
+                        Arguments = $"detach --hardware-id {fr.s_hardwareID}",
+                        RedirectStandardOutput = true,
+                        RedirectStandardError = true,
+                        UseShellExecute = false,
+                        CreateNoWindow = true
+                    };
 
-                using (Process process = new Process { StartInfo = startInfo })
-                {
-                    process.Start();
-                    process.WaitForExit();
-                   
+                    using (Process process = new Process { StartInfo = startInfo })
+                    {
+                        process.Start();
+                        process.WaitForExit();
+
+                    }
                 }
+                else
+                {
+                    SSHEnd();
+                }
+                
+                
             }
 
             catch (System.Exception ex)
@@ -133,13 +158,20 @@ namespace Photogrametry
             await Task.Delay(milliseconds);
         }
        //Read the number of photos on the camera to an integer for determining which files to download
-        private int GetCameraCount()
+        public int GetCameraCount()
         {
             try
             {
                 string Camera_Response;
-                WSLproc.StandardInput.WriteLine($"gphoto2 --num-files -f={fr.CameraFolder.Text}");
-                Camera_Response = WSLproc.StandardOutput.ReadLine();
+                if (fr.chk_UseWSL.Checked)
+                {
+                    WSLproc.StandardInput.WriteLine($"gphoto2 --num-files -f={fr.CameraFolder.Text}");
+                    Camera_Response = WSLproc.StandardOutput.ReadLine();
+                }
+                else
+                {
+                    Camera_Response = SSHWrite($"gphoto2 --num-files -f={fr.CameraFolder.Text}");
+                }
                 Match match = Regex.Match(Camera_Response, @"\d+$");
                 if (match.Success)
                 {
@@ -164,16 +196,81 @@ namespace Photogrametry
             startfiles = GetCameraCount();
             for (int i = 0; i < fr.i_Frames; i++)
             {
-                WSLproc.StandardInput.WriteLine($"gphoto2 --trigger-capture");
+                if (fr.chk_UseWSL.Checked)
+                {
+                    WSLproc.StandardInput.WriteLine($"gphoto2 --trigger-capture");
+                }
+                else
+                {
+                    SSHWrite($"gphoto2 --trigger-capture");
+                }
+                   
                 await WaitSeconds(fr.i_Interval);
                 //fr.LogMessage(WSLproc.StandardOutput.ReadToEnd());
             }
+            if (fr.chk_UseWSL.Checked)
+            {
+                WSLproc.StandardInput.WriteLine($"gphoto2 --get-file {startfiles + 1}-{startfiles + fr.i_Frames}");
+            }
+            else
+            {
+                SSHWrite($"gphoto2 --get-file {startfiles + 1}-{startfiles + fr.i_Frames}");
+            }
             
-            WSLproc.StandardInput.WriteLine($"gphoto2 --get-file {startfiles+1}-{startfiles+fr.i_Frames}" );
             
 
         }
-        
-        
+       public void SSHStart()
+        {
+            try
+            {
+                sshCLI = new SshClient("10.10.10.10", "ryanm", "Ryan1@3");
+
+                sshCLI.Connect();
+
+
+            }
+            catch (System.Exception ex)
+            {
+                MessageBox.Show("Unexpected error occurred: " + ex.Message);
+                fr.LogMessage(ex.Message + ex.Source + ex.StackTrace);
+            }
+        }
+
+        public string SSHWrite(string _cmd)
+        {
+            try
+            {
+                sshStream.WriteLine(_cmd);
+                string output = sshStream.Read();
+                Thread.Sleep(100);
+                Console.WriteLine(output);
+                return output;
+
+            }
+            catch (System.Exception ex)
+            {
+                MessageBox.Show("Unexpected error occurred: " + ex.Message);
+                fr.LogMessage(ex.Message + ex.Source + ex.StackTrace);
+                return "failed";
+            }
+        }
+        public void SSHEnd()
+        {
+            try
+            {
+                sshStream.Dispose();
+                sshCLI.Disconnect();
+                sshCLI.Dispose();
+
+
+            }
+            catch (System.Exception ex)
+            {
+                MessageBox.Show("Unexpected error occurred: " + ex.Message);
+                fr.LogMessage(ex.Message + ex.Source + ex.StackTrace);
+            }
+        }
+
+        }
     }
-}
